# backend/backend_app/static/openapi.yaml
openapi: "3.0.3"
info:
  title: API do Petshop
  description: API responsável pelo gerenciamento de usuários, pets e serviços
  version: 1.0.0
  contact: 
    name: Tiago F. Hinterholz
    url: https://github.com/tiagohinterholz/Petshop_API
    email: fh.tiago@gmail.com
servers:
  - url: http://localhost:5000
...
paths:
...
  /breeds:
    post:
      tags:
        - 'Breeds'
      operationId: cadastrarBreed
      summary: Cadastro de Raças de Pets
      security:
        - JWT: []
      requestBody:
        content:
          'application/json':
            schema:
              $ref: '#/components/schemas/BreedsRequest'
      responses:
        '201':
         description: "Breed cadastrada com sucesso"
         content:
          'application/json':
            schema:
              $ref: '#/components/schemas/BreedsResponse'
        '400':
          description: "Problema nos dados de cadastro"
          content:
            'application/json':
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '401':
          description: "Usuário não autenticado para esta operação"
          content:
            'application/json':
              schema:
                $ref: '#/components/schemas/ErrorResponse' 
        '500':
          description: "Erro inesperado ao cadastrar breed"
          content:
            'application/json':
              schema:
                $ref: '#/components/schemas/ErrorResponse'
    get:
      tags:
        - 'Breeds'
      operationId: listarContacts
      summary: Listagem de Raças de Pets
      security:
        - JWT: []
      responses:
        '200':
          description: "Listagem de raças realizada com sucesso"
          content:
            'application/json':
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/BreedsResponse'
        '401':
          description: 'Usuário não autorizado'
          content:
            'application/json':
              schema:
                $ref: '#/components/schemas/ErrorResponse'
  /breeds/{breed_id}:
    get:
      tags:
        - 'Breeds'
      operationId: detalharContact
      summary: "Detalhar Raça do Pet pelo breed_id"
      parameters:
        - name: breed_id
          in: path
          description: "ID da raça a ser detalhado"
          required: true
          schema:
            type: integer
      security:
        - JWT: []
      responses:
        '200':
          description: "Raça detalhada com sucesso"
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/BreedsResponse"
        '400':
          description: "Dados inseridos inválidos"
          content:
            'application/json':
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '401':
          description: "Usuario não autenticado"
          content:
            'application/json':
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '404':
          description: "Raça não encontrada"
          content:
            'application/json':
              schema:
                $ref: '#/components/schemas/ErrorResponse'   
    put:
      tags:
        - 'Breeds'
      summary: "Atualizar cadastro de Raças pelo breed_id"
      operationId: atualizarContact
      parameters:
        - name: breed_id
          in: path
          description: "ID da Raça do Pet a ser atualizada"
          required: true
          schema:
            type: integer
      security:
        - JWT: []
      requestBody:
        content:
          'application/json':
            schema:
              $ref: "#/components/schemas/BreedsRequest"
      responses:
        '200':
          description: "Raça atualizada com sucesso"
          content:
            'application/json':
              schema:
                $ref: '#/components/schemas/BreedsResponse'
        '400':
          description: "Dados inseridos inválidos"
          content:
            'application/json':
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '401':
          description: "Usuario não autenticado"
          content:
            'application/json':
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '404':
          description: "Raça não encontrada"
          content:
            'application/json':
              schema:
                $ref: '#/components/schemas/ErrorResponse' 
    delete:
      tags:
        - 'Breeds'
      operationId: deletarBreed
      summary: "Deletar Raça pelo breed_id"
      parameters:
        - name: breed_id
          in: path
          description: "ID da raça a ser deletada"
          required: true
          schema:
            type: integer
      security:
        - JWT: []
      responses:
        '200':
          description: "Raça deletada com sucesso"
        '401':
          description: "User não autenticado"
          content:
            'application/json':
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '404':
          description: "Raça não encontrada"
          content:
            'application/json':
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '500':
          description: "Erro ao excluir raça"
          content:
            'application/json':
              schema:
                $ref: '#/components/schemas/ErrorResponse'
  /appointments:
    post:
      tags:
        - 'Appointments'
      operationId: cadastrarBreed
      summary: Cadastro de Agendamento de Pets
      security:
        - JWT: []
      requestBody:
        content:
          'application/json':
            schema:
              $ref: '#/components/schemas/AppointmentsRequest'
      responses:
        '201':
         description: "Agendamento cadastrado com sucesso"
         content:
          'application/json':
            schema:
              $ref: '#/components/schemas/AppointmentsResponse'
        '400':
          description: "Problema nos dados de cadastro"
          content:
            'application/json':
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '401':
          description: "Usuário não autenticado para esta operação"
          content:
            'application/json':
              schema:
                $ref: '#/components/schemas/ErrorResponse' 
        '500':
          description: "Erro inesperado ao cadastrar appointment"
          content:
            'application/json':
              schema:
                $ref: '#/components/schemas/ErrorResponse'
    get:
      tags:
        - 'Appointments'
      operationId: listarAppointments
      summary: Listagem de Agendamentos de Pets
      security:
        - JWT: []
      responses:
        '200':
          description: "Listagem de agendamentos realizada com sucesso"
          content:
            'application/json':
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/AppointmentsResponse'
        '401':
          description: 'Usuário não autorizado'
          content:
            'application/json':
              schema:
                $ref: '#/components/schemas/ErrorResponse'
  /appointments/{appointment_id}:
    get:
      tags:
        - 'Appointments'
      operationId: detalharAppointment
      summary: "Detalhar Agendamento de Pet pelo appointment_id"
      parameters:
        - name: appointment_id
          in: path
          description: "ID do agendamento a ser detalhado"
          required: true
          schema:
            type: integer
      security:
        - JWT: []
      responses:
        '200':
          description: "Agendamento detalhada com sucesso"
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/AppointmentsResponse"
        '400':
          description: "Dados inseridos inválidos"
          content:
            'application/json':
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '401':
          description: "Usuario não autenticado"
          content:
            'application/json':
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '404':
          description: "Agendamento não encontrado"
          content:
            'application/json':
              schema:
                $ref: '#/components/schemas/ErrorResponse'   
    put:
      tags:
        - 'Appointments'
      summary: "Atualizar Agendamento de Pets pelo appointment_id"
      operationId: atualizarContact
      parameters:
        - name: breed_id
          in: path
          description: "ID do Agendamento do Pet a ser atualizado"
          required: true
          schema:
            type: integer
      security:
        - JWT: []
      requestBody:
        content:
          'application/json':
            schema:
              $ref: "#/components/schemas/AppointmentsRequest"
      responses:
        '200':
          description: "Agendamento atualizado com sucesso"
          content:
            'application/json':
              schema:
                $ref: '#/components/schemas/AppointmentsResponse'
        '400':
          description: "Dados inseridos inválidos"
          content:
            'application/json':
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '401':
          description: "Usuario não autenticado"
          content:
            'application/json':
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '404':
          description: "Agendamento não encontrado"
          content:
            'application/json':
              schema:
                $ref: '#/components/schemas/ErrorResponse' 
    delete:
      tags:
        - 'Appointments'
      operationId: deletarAppointment
      summary: "Deletar Agendamento de Pet pelo appointment_id"
      parameters:
        - name: appointment_id
          in: path
          description: "ID do Agendamento a ser deletado"
          required: true
          schema:
            type: integer
      security:
        - JWT: []
      responses:
        '200':
          description: "Agendamento deletado com sucesso"
        '401':
          description: "User não autenticado"
          content:
            'application/json':
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '404':
          description: "Agendamento não encontrado"
          content:
            'application/json':
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '500':
          description: "Erro ao excluir raça"
          content:
            'application/json':
              schema:
                $ref: '#/components/schemas/ErrorResponse'
  /addresses:
    post:
      tags:
        - 'Addresses'
      operationId: cadastrarAddress
      summary: Cadastro de Endereços
      security:
        - JWT: []
      requestBody:
        content:
          'application/json':
            schema:
              $ref: '#/components/schemas/AddressesRequest'
      responses:
        '201':
         description: "Endereço cadastrado com sucesso"
         content:
          'application/json':
            schema:
              $ref: '#/components/schemas/AddressesResponse'
        '400':
          description: "Problema nos dados de cadastro"
          content:
            'application/json':
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '401':
          description: "Usuário não autenticado para esta operação"
          content:
            'application/json':
              schema:
                $ref: '#/components/schemas/ErrorResponse' 
        '500':
          description: "Erro inesperado ao cadastrar address"
          content:
            'application/json':
              schema:
                $ref: '#/components/schemas/ErrorResponse'
    get:
      tags:
        - 'Addresses'
      operationId: listarAddress
      summary: Listagem de Endereços
      security:
        - JWT: []
      responses:
        '200':
          description: "Listagem de Endereços realizada com sucesso"
          content:
            'application/json':
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/AddressesResponse'
        '401':
          description: 'Usuário não autorizado'
          content:
            'application/json':
              schema:
                $ref: '#/components/schemas/ErrorResponse'
  /addresses/{address_id}:
    get:
      tags:
        - 'Addresses'
      operationId: detalharAddress
      summary: "Detalhar Endereço do Cliente pelo address_id"
      parameters:
        - name: address_id
          in: path
          description: "ID do endereço a ser detalhado"
          required: true
          schema:
            type: integer
      security:
        - JWT: []
      responses:
        '200':
          description: "Endereço detalhado com sucesso"
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/AddressesResponse"
        '400':
          description: "Dados inseridos inválidos"
          content:
            'application/json':
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '401':
          description: "Usuario não autenticado"
          content:
            'application/json':
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '404':
          description: "Endereço não encontrado"
          content:
            'application/json':
              schema:
                $ref: '#/components/schemas/ErrorResponse'   
    put:
      tags:
        - 'Addresses'
      summary: "Atualizar Endereço de Cliente pelo address_id"
      operationId: atualizarAddress
      parameters:
        - name: address_id
          in: path
          description: "ID do endereço a ser atualizado"
          required: true
          schema:
            type: integer
      security:
        - JWT: []
      requestBody:
        content:
          'application/json':
            schema:
              $ref: "#/components/schemas/AddressesRequest"
      responses:
        '200':
          description: "Endereço atualizado com sucesso"
          content:
            'application/json':
              schema:
                $ref: '#/components/schemas/AddressesResponse'
        '400':
          description: "Dados inseridos inválidos"
          content:
            'application/json':
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '401':
          description: "Usuario não autenticado"
          content:
            'application/json':
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '404':
          description: "Endereço não encontrado"
          content:
            'application/json':
              schema:
                $ref: '#/components/schemas/ErrorResponse' 
    delete:
      tags:
        - 'Addresses'
      operationId: deletarAddress
      summary: "Deletar Endereço pelo address_id"
      parameters:
        - name: address_id
          in: path
          description: "ID do Endereço a ser deletado"
          required: true
          schema:
            type: integer
      security:
        - JWT: []
      responses:
        '200':
          description: "Agendamento deletado com sucesso"
        '401':
          description: "User não autenticado"
          content:
            'application/json':
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '404':
          description: "Endereço não encontrado"
          content:
            'application/json':
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '500':
          description: "Erro ao excluir Endereço"
          content:
            'application/json':
              schema:
                $ref: '#/components/schemas/ErrorResponse'
components:
  schemas:
    ErrorResponse:
      type: object
      properties:
        error:
          type: string
...
    AddressesRequest:
      type: object
      required:
        - client_id
        - street
        - city
        - neighborhood
      properties:
        client_id:
          type: integer
          example: 1
        street:
          type: string
          maxLength: 50
          example: "Rua dos Pampas"
        city:
          type: string
          maxLength: 50
          example: "Brasília"
        neighborhood:
          type: string
          maxLength: 50
          example: "Bairro Biriquí"
        complement:
          type: string
          maxLength: 50
          exemplo: "Número 23"
    AddressesResponse:
      type: object
      properties:
        id:
         type: integer
        client_id:
          type: integer
        street:
          type: string
        city:
          type: string
        neighborhood:
          type: string
        complement:
          type: string
    AppointmentsRequest:
      type: object
      required:
        - pet_id
        - desc_appoint
        - price
        - date_appoint
      properties:
        pet_id:
          type: integer
          example: 1
        desc_appoint:
          type: string
          maxLength: 50
          example: "Banho e tosa"
        price:
          type: number
          format: float
          examploe: "550,00"
        date_appoint:
          type: string
          format: date
          example: "2025-05-31"
    AppointmentsResponse:
      type: object
      properties:
        id:
          type: integer
        pet_id:
          type: integer
        desc_appoint:
          type: string
        price:
          type: number
        date_appoint:
          type: string        
    BreedsRequest:
      type: object
      required:
        - description
      properties:
        description:
          type: string
          maxLength: 50
          example: "Gato Siamês / Cachorro Buldogue" 
    BreedsResponse:
      type: object
      properties:
        id:
         type: integer
        description:
          type: string
          maxLength: 50
...
  securitySchemes:
    JWT:
      type: http
      scheme: bearer  